+++ packages/block-editor/src/components/block-rename/empty-string.js
@@ -1,0 +1,1 @@
+export const emptyString = ( testString ) => testString?.trim()?.length === 0;
